refactor: move keys and storage rules to a dedicated folder
Referring to issue #123 but not closing it, as some refactoring will need to be done in the PR #121
in moving the email templates directory

#123
