+++ src/routes/storage/delete_file.ts
@@ -6,7 +6,7 @@ import { s3 } from '@shared/s3'
 import { verify } from '@shared/jwt'
 import { S3_BUCKET } from '@shared/config'
 
-import { storagePermission } from './rules'
+import { storagePermission } from '@custom/storage-rules'
 
 async function delete_file(req: Request, res: Response): Promise<unknown> {
   const key = req.params[0]
